% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posibatch.R
\name{posibatches}
\alias{posibatches}
\title{Adjust for positional effects and batch effects using posibatch in Illumina Beadchips}
\usage{
posibatches(dat, Sentrix, posi = TRUE, batch = TRUE, par.prior = TRUE,
  prior.plots = FALSE, mean.only.posi = FALSE, mean.only.batch = FALSE)
}
\arguments{
\item{dat}{Genomic measure matrix (dimensions probe x sample) - for example, expression matrix}

\item{Sentrix}{The list of position numbers and chip numbers - for example, sampleID, sampleNames (8963303124_R01C01) and batches}

\item{posi}{{Position covariate}}

\item{batch}{{Batch covariate}}

\item{par.prior}{(Optional) TRUE indicates parametric adjustments will be used, FALSE indicates non-parametric adjustments will be used}

\item{prior.plots}{(Optional) TRUE give prior plots with black as a kernel estimate of the empirical batch effect density and red as the parametric}

\item{mean.only.posi}{(Optional) FALSE If TRUE posibatch only corrects the mean of the positional effect (no scale adjustment). If one position has only one sample, setting mean.only=TRUE.}

\item{mean.only.batch}{(Optional) FALSE If TRUE posibatch only corrects the mean of the batch effect (no scale adjustment). If one batch has only one sample, setting mean.only=TRUE.}

\item{mod}{Model matrix for outcome of interest and other covariates besides batch and position}
}
\value{
data A probe x sample genomic measure matrix, adjusted for positional effects and batch effects.
}
\description{
posibatch allows users to adjust for positional effects and batch effects in Illumina Beadchips. Positional effects means the same sample in
 different physical positions on the array could be measured as different methylation or expression levels, and batch effects. Batch effects
 are sub-groups of measurements that have qualitatively different behaviour across conditions and are unrelated to the biological or
 scientific variables in a study. It uses either parametric or non-parametric empirical Bayes frameworks for adjusting data for positional
 effects and batch effects. Users are returned an expression or methylation matrix that has been corrected for position effects and batch effects. 
 The input data are assumed to be cleaned and normalized before batch effect removal.
}
